Feature user

+++ src/Controller/Front/TipController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Front;
 
 use App\Entity\Tip;
 use App\Form\TipType;
@@ -12,18 +12,11 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Routing\Requirement\Requirement;
 
-class AstuceController extends AbstractController
+class TipController extends AbstractController
 {
-    #[Route('/', name: 'app_astuce')]
-    public function index(TipRepository $tipRepository): Response
-    {
-        return $this->render('astuce/index.html.twig', [
-            'tips' => $tipRepository->findAll(),
-        ]);
-    }
-
-    #[Route('/astuce/{id}', name: 'app_astuce_show', methods: ['GET'])]
+    #[Route('/tip/{id}', name: 'app_front_tip_show', methods: ['GET'], requirements: ['id' => Requirement::DIGITS])]
     public function show(int $id, TipRepository $tipRepository): Response
     {
         $result = $tipRepository->findDetailOfOneTip($id);
@@ -32,34 +25,41 @@ class AstuceController extends AbstractController
             throw $this->createNotFoundException('Cette astuce n\'existe pas.');
         }
 
-        return $this->render('astuce/show.html.twig', [
+        return $this->render('front/tip/show.html.twig', [
             'tip' => $result['tip'],
             'ingredientQuantities' => $result['ingredientQuantities'],
         ]);
     }
 
-    #[Route('/new', name: 'app_astuce_new')]
+    #[Route('/tip/new', name: 'app_front_tip_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, TipService $tipService): Response
     {
+        // Check if the user is authenticated
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_REMEMBERED');
+
         $tip = new Tip();
         $form = $this->createForm(TipType::class, $tip);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            // Get the current authenticated user
+            $user = $this->getUser();
+
             // Prepare the tip for persistence (ex. setting slug, status, etc.)
-            $tipService->prepareTipForPersistence($tip);
+            $tipService->prepareTipForPersistence($tip, $user);
             $entityManager->persist($tip);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_astuce');
+            return $this->redirectToRoute('app_home');
         }
 
-        return $this->render('astuce/new.html.twig', [
+        return $this->render('front/tip/new.html.twig', [
             'form' => $form,
         ]);
     }
 
-    #[Route('/astuce/{id}/edit', name: 'app_astuce_edit', methods: ['GET', 'POST'])]
+    #[Route('/tip/{id}/edit', name: 'app_front_tip_edit', methods: ['GET', 'POST'], requirements: ['id' => Requirement::DIGITS])]
     public function edit(Request $request, Tip $tip, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(TipType::class, $tip);
@@ -69,21 +69,23 @@ class AstuceController extends AbstractController
             $tip->setUpdatedAt(new DateTimeImmutable());
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_astuce', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('astuce/edit.html.twig', [
+        return $this->render('front/tip/edit.html.twig', [
             'tip' => $tip,
             'form' => $form,
         ]);
     }
 
-    #[Route('/astuce/{id}/delete', name: 'app_astuce_delete')]
-    public function delete(Tip $tip, EntityManagerInterface $entityManager): Response
+    #[Route('/tip/{id}', name: 'app_front_tip_delete', methods: ['POST'], requirements: ['id' => Requirement::DIGITS])]
+    public function delete(Request $request, Tip $tip, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($tip);
-        $entityManager->flush();
+        if ($this->isCsrfTokenValid('delete' . $tip->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($tip);
+            $entityManager->flush();
+        }
 
-        return $this->redirectToRoute('app_astuce', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
     }
 }
